make `buildSeasonalScoreboardEmbed()` and `buildOverallScoreboardEmbed()` into `Company` methods
