+++ source/commands/scoreboard.js
@@ -1,6 +1,5 @@
 const { InteractionContextType, MessageFlags } = require('discord.js');
 const { CommandWrapper } = require('../classes');
-const { buildSeasonalScoreboardEmbed, buildOverallScoreboardEmbed } = require('../util/embedUtil');
 
 /** @type {typeof import("../logic")} */
 let logicLayer;
@@ -8,22 +7,16 @@ let logicLayer;
 const mainId = "scoreboard";
 module.exports = new CommandWrapper(mainId, "View the XP scoreboard", null, false, [InteractionContextType.BotDM, InteractionContextType.Guild, InteractionContextType.PrivateChannel], 3000,
 	/** View the XP scoreboard */
-	(interaction, database, runMode) => {
-		if (interaction.options.getString("scoreboard-type") === "season") {
-			buildSeasonalScoreboardEmbed(interaction.guild, logicLayer).then(embed => {
-				interaction.reply({
-					embeds: [embed],
-					flags: [MessageFlags.Ephemeral]
-				});
-			})
-		} else {
-			buildOverallScoreboardEmbed(interaction.guild, database, logicLayer).then(embed => {
-				interaction.reply({
-					embeds: [embed],
-					flags: [MessageFlags.Ephemeral]
-				});
-			})
-		}
+	async (interaction, database, runMode) => {
+		const [company] = await logicLayer.companies.findOrCreateCompany(interaction.guild.id);
+		interaction.reply({
+			embeds: [
+				interaction.options.getString("scoreboard-type") === "season" ?
+					await company.seasonalScoreboardEmbed(interaction.guild, logicLayer) :
+					await company.overallScoreboardEmbed(interaction.guild, logicLayer)
+			],
+			flags: [MessageFlags.Ephemeral]
+		});
 	}
 ).setOptions(
 	{
